Refactor delete function to fix redirect bug
The delete function in main.py has been refactored to fix a bug where the redirect was not correctly pointing to the index.html page. The redirect URL has been updated to 'index.html' instead of 'html'. This change ensures that the user is redirected to the correct page after deleting a file.