Prototying (#50)
* chore: Update username container style in style.css and script.js

The username container style in the style.css file and script.js file has been updated to include the ".username_container" class. This change allows for proper display and alignment of the username container in the application.

* chore: Update username container style in style.css and script.js

* chore: Update FLASK_RUN_HOST to localhost in Dockerfile

The FLASK_RUN_HOST environment variable in the Dockerfile has been updated to "localhost" instead of "0.0.0.0". This change ensures that the Flask application runs on the local host only, improving security and preventing access from external networks.

* chore: Update FLASK_RUN_HOST to 0.0.0.0 in Dockerfile

* chore: Update FLASK_RUN_HOST to 0.0.0.0 in Dockerfile

* chore: Refactor Bio.load_json and Bio.qr_code functions

The Bio.load_json and Bio.qr_code functions in the Bioprocess module have been refactored to remove the unused 'cookie' parameter. This change improves code readability and removes unnecessary complexity.

* Refactor Bio.qr_code function to remove unused 'cookie' parameter

* Refactor Bioprocess module to remove unused 'cookie' parameter

* Refactor Blast_web module to remove unused code and improve performance

The Blast_web module has been refactored to remove unused code and improve performance. The web_blast function has been removed as it was not being used. Additionally, the blast.xml and blast.txt files are no longer generated, resulting in faster execution and reduced disk space usage.

* chore: Update results.html to pass username as a parameter

* Refactor delete function to fix redirect bug

The delete function in main.py has been refactored to fix a bug where the redirect was not correctly pointing to the index.html page. The redirect URL has been updated to 'index.html' instead of 'html'. This change ensures that the user is redirected to the correct page after deleting a file.

* Refactor Dockerfile to use Python 3.12.3-slim

The Dockerfile has been updated to use the official Python 3.12.3-slim runtime instead of Python 3.11.3-slim. This change ensures that the application is built and run with the latest stable version of Python, improving compatibility and security.

* chore: Remove unused code and update QR code generation

The code changes in this commit remove unused code and update the QR code generation in the Flask application. The unused 'cookie' parameter has been removed from the Bio.qr_code function in the Bioprocess module. Additionally, the QR code URL has been updated to use the 'api/results' endpoint instead of the previous URL. This change improves code readability and ensures the QR code is generated correctly.

* Remove unused code and update QR code generation

* Refactor Python version to 3.12.3 in README.md